Extract raster modules so lint stays under module-size ceilings
The CI module-size gate is shrink-only. Move GAME.json image baking,
source-byte helpers, and matching tests out of the files this PR grew.
