+++ apps/api/src/catalog/github-client.ts
@@ -20,23 +20,13 @@ import {
   DELIVERY_FIXED_FILES,
   GAME_KIT_MODULES,
   GAME_KIT_VERTICAL_ENTRIES,
-  IMAGES_CONTRACT,
   MAX_SOURCE_GRAPH_MODULES,
   MUSIC_CONTRACT,
   SOURCE_GRAPH_BUDGET_BYTES,
   type GameKitModuleName,
 } from '../platform/games-repo-contract.js';
-import {
-  assertImageFileSize,
-  assertImageSignature,
-  decodeRasterSourceContent,
-  encodeRasterSourceContent,
-  imageLoaderBootJs,
-  imageLoaderHtml,
-  mimeForImagePath,
-  parseGameImages,
-  type ImageManifest,
-} from './raster-assets.js';
+import { appendDeclaredImageSources, bakeGameImageAssets, resolveGameImageBytes } from './bake-game-images.js';
+import { parseGameManifest } from './parse-game-manifest.js';
 import { isRateLimitResponse } from '../platform/github-rate-limit.js';
 import {
   generateIndexHtml,
@@ -214,96 +204,8 @@ export function resolveGameTypeScriptPath(resolveDir: string, specifier: string)
   return `${resolvedPath}.ts`;
 }
 
-interface GameManifest {
-  engine?: { modules?: unknown };
-  audio?: { sounds?: unknown; music?: unknown; musicTracks?: unknown };
-  images?: unknown;
-}
-
 interface SourcedAudioCatalog {
   sounds?: Record<string, { mime?: unknown }>;
-}
-
-interface ParsedGameManifest {
-  modules: GameKitModuleName[];
-  sounds: string[];
-  /**
-   * Selected BGM track id from GAME.json (`audio.music` string). Null when the
-   * audio module is off. Matches games-repo `tools/lib/assemble.ts`.
-   */
-  music: string | null;
-  /**
-   * Extra BGM ids from GAME.json (`audio.musicTracks`), embedded alongside `music` so a
-   * game can change score mid-round without a fetch. Empty for almost every game.
-   */
-  musicTracks: string[];
-  // GAME.json images map; empty when the game has no rasters.
-  images: ImageManifest;
-}
-
-function isKebabCaseName(value: unknown): value is string {
-  return typeof value === 'string' && /^[a-z0-9][a-z0-9-]*$/.test(value);
-}
-
-function parseGameManifest(source: string): ParsedGameManifest {
-  const manifest = JSON.parse(source) as GameManifest;
-  const modules = manifest.engine?.modules;
-  if (
-    !Array.isArray(modules) ||
-    modules.some(
-      (moduleName) =>
-        typeof moduleName !== 'string' || !GAME_KIT_MODULES.some((allowedModule) => allowedModule === moduleName),
-    )
-  ) {
-    throw new Error('game manifest contains invalid engine modules');
-  }
-
-  const expectedOrder = GAME_KIT_MODULES.filter((moduleName) => modules.includes(moduleName));
-  if (new Set(modules).size !== modules.length || modules.join(',') !== expectedOrder.join(',')) {
-    throw new Error('game manifest engine modules are duplicated or out of order');
-  }
-
-  const images = parseGameImages(manifest.images);
-
-  if (!modules.includes('audio')) {
-    return { modules: modules as GameKitModuleName[], sounds: [], music: null, musicTracks: [], images };
-  }
-
-  const sounds = manifest.audio?.sounds;
-  if (
-    !Array.isArray(sounds) ||
-    sounds.length === 0 ||
-    new Set(sounds).size !== sounds.length ||
-    !sounds.every(isKebabCaseName)
-  ) {
-    throw new Error('game manifest contains invalid audio sounds');
-  }
-
-  // Games-repo assemble: `audio.music` is a single track name string. Injected as
-  // `window.__GAME_AUDIO_MUSIC__ = "<name>"`, and it is the track that autoplays.
-  const music = manifest.audio?.music;
-  if (!isKebabCaseName(music)) {
-    throw new Error('game manifest contains invalid audio music');
-  }
-
-  // `audio.musicTracks` is optional. Mirrors games-repo validate Check 3: non-empty when
-  // present, kebab-case names, no duplicates, and never a repeat of `audio.music`.
-  const rawTracks = manifest.audio?.musicTracks;
-  let musicTracks: string[] = [];
-  if (rawTracks !== undefined) {
-    if (
-      !Array.isArray(rawTracks) ||
-      rawTracks.length === 0 ||
-      new Set(rawTracks).size !== rawTracks.length ||
-      !rawTracks.every(isKebabCaseName) ||
-      rawTracks.includes(music)
-    ) {
-      throw new Error('game manifest contains invalid audio musicTracks');
-    }
-    musicTracks = rawTracks;
-  }
-
-  return { modules: modules as GameKitModuleName[], sounds, music, musicTracks, images };
 }
 
 /**
@@ -1409,21 +1311,9 @@ export function createGitHubClient(options: GitHubClientOptions): GitHubClient {
       }
       const manifestSource = sources['GAME.json'];
       if (manifestSource) {
-        let images: ImageManifest;
-        try {
-          images = parseGameImages((JSON.parse(manifestSource) as { images?: unknown }).images);
-        } catch {
-          images = {};
-        }
-        for (const [name, relPath] of Object.entries(images)) {
-          const bytes = await readRawBytes(`games/${slug}/${relPath}`, ref);
-          if (!bytes) {
-            throw new Error(`game image "${name}" not found: ${relPath}`);
-          }
-          assertImageFileSize(name, bytes.byteLength);
-          assertImageSignature(name, relPath, bytes);
-          sources[relPath] = encodeRasterSourceContent(bytes);
-        }
+        await appendDeclaredImageSources(sources, manifestSource, (relPath) =>
+          readRawBytes(`games/${slug}/${relPath}`, ref),
+        );
         // Every game relies on this — neither file is ever committed anymore.
         if (!sources['index.html']?.trim()) {
           const title = sources['SPEC.md'] ? parseSpecTitle(sources['SPEC.md']) : null;
@@ -1471,19 +1361,6 @@ export function createGitHubClient(options: GitHubClientOptions): GitHubClient {
         if (overrides && Object.hasOwn(overrides, relative)) return overrides[relative];
         if (options?.noRefFallback) return null;
         return await readRawFile(`games/${slug}/${relative}`, ref);
-      };
-
-      const gameImageBytes = async (relative: string, name: string): Promise<Uint8Array | null> => {
-        if (overrides && Object.hasOwn(overrides, relative)) {
-          try {
-            return decodeRasterSourceContent(relative, overrides[relative]);
-          } catch (error) {
-            const detail = error instanceof Error ? error.message : 'invalid raster';
-            throw new Error(`game image "${name}" (${relative}): ${detail}`, { cause: error });
-          }
-        }
-        if (options?.noRefFallback) return null;
-        return await readRawBytes(`games/${slug}/${relative}`, ref);
       };
 
       const startedAt = Date.now();
@@ -1622,22 +1499,15 @@ export function createGitHubClient(options: GitHubClientOptions): GitHubClient {
       }
 
       let loaderHtml = '';
-      const imageNames = Object.keys(manifest.images);
-      if (imageNames.length > 0) {
-        const imageAssets: Record<string, string> = {};
-        for (const name of imageNames) {
-          const relPath = manifest.images[name];
-          const bytes = await gameImageBytes(relPath, name);
-          if (!bytes) {
-            throw new Error(`game image "${name}" not found: ${relPath}`);
-          }
-          assertImageFileSize(name, bytes.byteLength);
-          assertImageSignature(name, relPath, bytes);
-          imageAssets[name] = `data:${mimeForImagePath(relPath)};base64,${Buffer.from(bytes).toString('base64')}`;
-        }
-        assetChunks.push(`window.${IMAGES_CONTRACT.windowAssetsName} = Object.freeze(${JSON.stringify(imageAssets)});`);
-        assetChunks.push(imageLoaderBootJs(imageNames));
-        loaderHtml = imageLoaderHtml();
+      const bakedImages = await bakeGameImageAssets(manifest.images, (relPath, name) =>
+        resolveGameImageBytes(relPath, name, overrides, options?.noRefFallback, () =>
+          readRawBytes(`games/${slug}/${relPath}`, ref),
+        ),
+      );
+      if (bakedImages) {
+        assetChunks.push(bakedImages.assetChunk);
+        assetChunks.push(bakedImages.bootJs);
+        loaderHtml = bakedImages.loaderHtml;
       }
 
       const assetsJs = assetChunks.length > 0 ? `${assetChunks.join('\n')}\n` : '';
