+++ apps/api/src/platform/games-repo-contract.test.ts
@@ -18,10 +18,10 @@ import {
   MAX_PROJECT_BYTES,
   MAX_SOURCE_GRAPH_MODULES,
   MUSIC_CONTRACT,
-  RASTER_ASSET_BUDGET_BYTES,
   SOURCE_GRAPH_BUDGET_BYTES,
   stripLeadingDocComment,
 } from './games-repo-contract.js';
+import { RASTER_ASSET_BUDGET_BYTES } from './raster-contract.js';
 import { MAX_PROJECT_BYTES as ASSEMBLE_MAX } from './assemble.js';
 import { ALLOWED_SOURCE_FILES, MAX_UPLOAD_BYTES, MAX_UPLOAD_FILES } from '../delivery/games-store.js';
 
@@ -275,30 +275,6 @@ describe('games-repo source extractors', () => {
       expect(GAME_KIT_MODULES).toContain(name);
       expect(entry).toMatch(/^shared\/(verticals|modules)\/[a-z0-9-]+\/index\.ts$/);
     }
-  });
-
-  it('evaluates MAX_BUNDLE_BYTES from the single platform ceiling', () => {
-    // Games-repo #281 collapsed the per-feature ledger into one GAMEKIT_PLATFORM_BYTES.
-    // The extractor still has to resolve `BUDGET + PLATFORM` (and `a + b` / `a * b`
-    // forms inside those consts), not only bare literals.
-    const source = `
-      const GAME_BUDGET_BYTES = 936 * 1024;
-      const GAMEKIT_PLATFORM_BYTES = 410_000;
-      const RASTER_ASSET_BUDGET_BYTES = 3 * 1024 * 1024;
-      const MAX_BUNDLE_BYTES = GAME_BUDGET_BYTES + GAMEKIT_PLATFORM_BYTES + RASTER_ASSET_BUDGET_BYTES;
-    `;
-    expect(extractMaxBundleBytes(source)).toBe(MAX_PROJECT_BYTES);
-  });
-
-  it('still evaluates a + b allowance expressions when a tip uses them', () => {
-    const source = `
-      const GAME_BUDGET_BYTES = 936 * 1024;
-      const GAMEKIT_TOUCH_BYTES = 7_501 + 5_560;
-      const GAMEKIT_PLATFORM_BYTES = GAMEKIT_TOUCH_BYTES + 396_939;
-      const RASTER_ASSET_BUDGET_BYTES = 3 * 1024 * 1024;
-      const MAX_BUNDLE_BYTES = GAME_BUDGET_BYTES + GAMEKIT_PLATFORM_BYTES + RASTER_ASSET_BUDGET_BYTES;
-    `;
-    expect(extractMaxBundleBytes(source)).toBe(MAX_PROJECT_BYTES);
   });
 
   it('evaluates a numeric MAX_BUNDLE_BYTES literal', () => {
