+++ eslint-rules/module-boundary-map.mjs
@@ -111,6 +111,8 @@ const FILE_BUCKET = {
   assemble: 'platform',
   'games-repo-archive': 'platform',
   'games-repo-contract': 'platform',
+  'games-repo-bundle-bytes': 'platform',
+  'raster-contract': 'platform',
   'kit-registry': 'platform',
   'kit-window': 'platform',
   'round-base-version': 'platform',
@@ -260,6 +262,11 @@ const FILE_BUCKET = {
   'staged-preview': 'delivery',
   'stage-hints': 'delivery',
   'games-store': 'delivery',
+  'games-store-raster': 'delivery',
+  'delivery-path-guard': 'delivery',
+  'source-file-bytes': 'delivery',
+  'gate-materialize': 'delivery',
+  'gate-runner-raster': 'delivery',
   'workspace-archive': 'delivery',
   'build-transcript': 'delivery',
   'build-changelog': 'delivery',
@@ -292,6 +299,9 @@ const FILE_BUCKET = {
   'catalog-search-routes': 'catalog',
   'embedding-service': 'catalog',
   'github-client': 'catalog',
+  'github-client-images': 'catalog',
+  'bake-game-images': 'catalog',
+  'parse-game-manifest': 'catalog',
   'raster-assets': 'catalog',
   'game-snapshot': 'catalog',
   'game-snapshot-publish': 'catalog',
